+++ src/index.js
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom';
 import 'normalize.css';
 import './index.css';
 import App from './App';
-import AppWrapper from './configuration/AppWrapper'
-import AppErrorBoundary from './configuration/AppErrorBoundary'
+import AppWrapper from 'src/providers/AppWrapper'
+import AppErrorBoundary from 'src/providers/AppErrorBoundary'
 import { ConfigurationWrapper } from 'src/features/remoteConfig/'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { queryClient } from 'src/lib/react-query'
